refactor settings urls
